move ClientChat and ClientIndependentChatRoom into chat folder
