+++ component/client/chat/ClientIndependentChatRoom.tsx
